requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==1.2.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
coloredlogs==15.0.1
cytoolz==0.11.2
diskcache==5.4.0
ecdsa==0.17.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
humanfriendly==10.0
humanize==3.13.1
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
more-itertools==8.12.0
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
py-bip39-bindings==0.1.9
py-ed25519-bindings==1.0.1
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.1.4
pycparser==2.21
pycryptodome==3.12.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.0
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rlp==2.0.1
scalecodec==1.0.45
six==1.16.0
substrate-interface==1.3.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
varint==1.0.2
web3==5.29.0
websocket-client==1.3.1
websockets==9.1
xxhash==2.0.2
yarl==1.7.2
zipp==3.7.0
